refactored the code to extract a util function
