+++ frontend/src/metabase/visualizations/lib/save-dashboard-pdf.ts
@@ -1,6 +1,7 @@
 import { t } from "ttag";
 
 import { isStorybookActive } from "metabase/env";
+import { openImageBlobOnStorybook } from "metabase/lib/loki-utils";
 import type { Dashboard } from "metabase-types/api";
 
 import { SAVING_DOM_IMAGE_CLASS } from "./save-chart-image";
@@ -51,19 +52,7 @@ export const saveDashboardPdf = async (
     // so we can test the export result with loki
     image.toBlob(blob => {
       if (blob) {
-        const imgElement = document.createElement("img");
-        imgElement.src = URL.createObjectURL(blob);
-        // scale to /2 to compensate `scale:2` in html2canvas
-        imgElement.width = image.width / 2;
-        imgElement.height = image.height / 2;
-
-        const root: HTMLElement = window.document.querySelector("#root")!;
-        const imageDownloaded = document.createElement("div");
-        imageDownloaded.setAttribute("data-testid", "image-downloaded");
-        root.replaceChildren(imgElement);
-        root.appendChild(imageDownloaded);
-
-        window.document.body.style.height = "initial";
+        openImageBlobOnStorybook({ canvas: image, blob });
       }
     });
   } else {
